Minor bugfixes

+++ src/main/scanner/helpers/math_helpers.hpp
@@ -12,9 +12,9 @@ bool intersection_line_segment(const Eigen::Vector2d& a1, const Eigen::Vector2d&
 //TODO: improve by accounting for parallel lines
 //bool intersects_line_segment(const Eigen::Vector2f& a1, const Eigen::Vector2f& b1, const Eigen::Vector2f& a2, const Eigen::Vector2d& b2);
 bool intersection_line(const Eigen::Vector2d& a1, const Eigen::Vector2d& b1, const Eigen::Vector2d& a2, const Eigen::Vector2d& b2, Eigen::Vector2d& intersection);
-Eigen::Vector3d proj_a2b3D(const Eigen::Vector3f& a, const Eigen::Vector3f& b, float b_squared_norm);
+Eigen::Vector3d proj_a2b3D(const Eigen::Vector3f& a, const Eigen::Vector3f& b, double b_squared_norm);
 Eigen::Vector3d proj_a2b3D(const Eigen::Vector3f& a, const Eigen::Vector3f& b);
-float cross_ratio(float A, float B, float C, float a, float b, float c, float q);
+double cross_ratio(double A, double B, double C, double a, double b, double c, double q);
 bool inside_polygon(const Eigen::Vector2f& a, const std::vector<line_segment>& sides);
 
 }
